+++ usermgnt/data/default_data_adapter.py
@@ -13,6 +13,7 @@ Created on 17 june 2019
 
 
 from usermgnt.common.logs import LOG
+from usermgnt.data.standalone import data as data
 
 
 # Data adapter class
@@ -20,20 +21,7 @@ class StandaloneDataAdapter:
 
     # FUNCTION: get_current_device_id
     def get_current_device_id(self):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_current_device_id] not implemented")
-        return None
-
-
-    # FUNCTION: get_current_device_ip
-    def get_current_device_ip(self):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_current_device_ip] not implemented")
-        return None
-
-
-    # FUNCTION: get_leader_device_ip
-    def get_leader_device_ip(self):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_leader_device_ip] not implemented")
-        return None
+        return data.get_current_device_id()
 
 
     # FUNCTION: get_agent_info
@@ -62,38 +50,32 @@ class StandaloneDataAdapter:
 
     # FUNCTION: get_user_profile_by_id
     def get_sharing_model_by_id(self, sharing_model_id):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_sharing_model_by_id] not implemented")
-        return None
+        return data.get_sharing_model_by_id(sharing_model_id)
 
 
     # Get shared resources
     def get_sharing_model(self, device_id):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_sharing_model] not implemented")
-        return None
+        return data.get_sharing_model(device_id)
 
 
     # Initializes shared resources values
-    def init_sharing_model(self, data):
-        LOG.warning("[usermgnt.data.default_data_adapter] [init_sharing_model] not implemented")
-        return None
+    def init_sharing_model(self, d):
+        return data.init_sharing_model(d)
 
 
     # Updates shared resources values
-    def update_sharing_model_by_id(self, sharing_model_id, data):
-        LOG.warning("[usermgnt.data.default_data_adapter] [update_sharing_model_by_id] not implemented")
-        return None
+    def update_sharing_model_by_id(self, sharing_model_id, d):
+        return data.update_sharing_model_by_id(sharing_model_id, d)
 
 
     # delete_sharing_model_by_id: Deletes  shared resources values
     def delete_sharing_model_by_id(self, sharing_model_id):
-        LOG.warning("[usermgnt.data.default_data_adapter] [delete_sharing_model_by_id] not implemented")
-        return None
+        return data.delete_sharing_model_by_id(sharing_model_id)
 
 
     # FUNCTION: get_current_sharing_model: Get current SHARING-MODEL
     def get_current_sharing_model(self):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_current_sharing_model] not implemented")
-        return None
+        return data.get_current_sharing_model()
 
 
     ###############################################################################
@@ -101,44 +83,37 @@ class StandaloneDataAdapter:
 
     # get_user_profile_by_id
     def get_user_profile_by_id(self, profile_id):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_user_profile_by_id] not implemented")
-        return None
+        return data.get_user_profile_by_id(profile_id)
 
 
     # get_user_profile: Get user profile
     def get_user_profile(self, device_id):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_user_profile] not implemented")
-        return None
+        return data.get_user_profile(device_id)
 
 
     # update_user_profile_by_id: Updates a profile
-    def update_user_profile_by_id(self, profile_id, data):
-        LOG.warning("[usermgnt.data.default_data_adapter] [update_user_profile_by_id] not implemented")
-        return None
+    def update_user_profile_by_id(self, profile_id, d):
+        return data.update_user_profile_by_id(profile_id, d)
 
 
     # Deletes users profile
     def delete_user_profile_by_id(self, profile_id):
-        LOG.warning("[usermgnt.data.default_data_adapter] [delete_user_profile_by_id] not implemented")
-        return None
+        return data.delete_user_profile_by_id(profile_id)
 
 
     # Initializes users profile
-    def register_user(self, data):
-        LOG.warning("[usermgnt.data.default_data_adapter] [register_user] not implemented")
-        return None
+    def register_user(self, d):
+        return data.register_user(d)
 
 
     # setAPPS_RUNNING
     def setAPPS_RUNNING(self, apps=0):
-        LOG.warning("[usermgnt.data.default_data_adapter] [setAPPS_RUNNING] not implemented")
-        return None
+        return data.setAPPS_RUNNING(apps)
 
 
     # FUNCTION: get_current_user_profile: Get Current USER-PROFILE
     def get_current_user_profile(self):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_current_user_profile] not implemented")
-        return None
+        return data.get_current_user_profile()
 
 
     ###############################################################################
@@ -147,14 +122,12 @@ class StandaloneDataAdapter:
 
     # FUNCTION: get_total_services_running: Get services running
     def get_total_services_running(self):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_total_services_running] not implemented")
-        return None
+        return data.get_total_services_running()
 
 
     # FUNCTION: get_power: Get battery level from DEVICE_DYNAMIC
     def get_power(self):
-        LOG.warning("[usermgnt.data.default_data_adapter] [get_power] not implemented")
-        return None
+        return data.get_power()
 
 
     ###############################################################################
@@ -163,11 +136,9 @@ class StandaloneDataAdapter:
 
     # FUNCTION: save_device_id
     def save_device_id(self, device_id):
-        LOG.warning("[usermgnt.data.default_data_adapter] [save_device_id] not implemented")
-        return None
+        return data.save_device_id(device_id)
 
 
     # FUNCTION: read_device_id
     def read_device_id(self):
-        LOG.warning("[usermgnt.data.default_data_adapter] [read_device_id] not implemented")
-        return None
+        return data.read_device_id()