+++ CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.7] - 2019-06-17
+## [1.2.8] - 2019-06-17
 ### Added
 - Changelog file added to project
 - data adpater for mF2C and standalone modes
+- standalone mode implemented
 
 ### Changed
 - packages structure updated
 
 ### Fixed
-- ...
+- deleted user method